Decouple from wordpress
