+++ assets/src/theme/renderSelectedVars.js
@@ -1,8 +1,9 @@
 import { VarPicker } from './VarPicker';
 
-export const renderSelectedVars = (rootElement, cssVars = [], lastTarget, groups, rawGroups, allVars) => {
+export const renderSelectedVars = (rootElement, cssVars = [], lastTarget, groups, rawGroups, allVars, config) => {
   wp.element.render(
     <VarPicker
+      config={ config }
       initialOpen={ false }
       selectedVars={ cssVars }
       groups={ groups }
