+++ assets/src/themeEditorIndex.js
@@ -1,0 +1,8 @@
+import { setupThemeEditor } from './theme/initializeThemeEditor';
+import { p4ServerThemes } from './theme/p4ServerThemes';
+
+document.addEventListener('DOMContentLoaded', async () => {
+  setupThemeEditor({
+    serverThemes: p4ServerThemes,
+  });
+});
