+++ webpack.config.js
@@ -32,7 +32,7 @@ module.exports = {
     style: './assets/src/styles/style.scss',
     editorStyle: './assets/src/styles/editorStyle.scss',
     lightbox: './assets/src/styles/lightbox.scss',
-    themeEditor: './assets/src/theme/initializeThemeEditor.js',
+    themeEditor: './assets/src/themeEditorIndex.js',
     themeEditorStyle: './assets/src/styles/themeEditor.scss',
     theme_antarctic: './assets/src/styles/theme_antarctic.scss',
     theme_arctic: './assets/src/styles/theme_arctic.scss',
